+++ Components/MapPlayer.js
@@ -40,6 +40,12 @@ class MapPlayer {
         this.maxRooms = 0
         /**@type {[MapPlayer[], import("./Room").default][]} */
         this.roomsData = []
+
+        this.skyblockLevel = null
+        this.dungeonClass = null
+        this.classLevel = null
+
+        this.playerColor = [0, 0, 0, 255]
     }
 
     get username() {
@@ -53,6 +59,49 @@ class MapPlayer {
         this.uuid = undefined
         this.checkUpdateUUID()
     }
+
+    updateDungeonClass() {
+        // for everything in tab
+        for (let entry of TabList.getNames()) {
+            // remove everything thats not a letter space or number and get rid of color codes
+            entry = ChatLib.removeFormatting(entry).replace('[YOUTUBE] ', '').replace('[ADMIN] ', '').replace(/[^A-Za-z0-9 _]/gi, '').replace('  ', ' ')
+            // if players username is it it 
+            if (entry.includes(this.username)){
+                // pull out the cool stuff and then leave
+                this.skyblockLevel = entry.split(' ')[0]
+                this.dungeonClass = entry.split(' ')[2]
+                this.classLevel = entry.split(' ')[3] // This can be simplified if all the class colors become an object, but thats weird so i didnt do that...
+                return this
+            }
+          }
+          return this
+    }
+
+    updatePlayerColor() {
+        if (settings.settings.headBorder == "single"){
+            this.playerColor = settings.settings.singleBorderColor
+            return this
+        }
+        switch(this.dungeonClass) {
+            case "Healer":
+                this.playerColor = settings.settings.healerColor ?? [240, 70, 240, 255]
+                break;
+            case "Mage":
+                this.playerColor = settings.settings.mageColor ?? [70, 210, 210, 255]
+                break;
+            case "Berserk":
+                this.playerColor = settings.settings.bersColor ?? [255, 0, 0, 255]
+                break;
+            case "Archer":
+                this.playerColor = settings.settings.archerColor ?? [30, 170, 50, 255]
+                break;
+            case "Tank":
+                this.playerColor = settings.settings.tankColor ?? [150, 150, 150, 255]
+                break;
+        }
+        return this
+    }
+    
 
     checkUpdateUUID() {
         if (this.uuid) return
@@ -119,9 +168,12 @@ class MapPlayer {
         Renderer.translate(x + w / 2, y + h / 2, 50)
         // Renderer.translate(x + (this.location.worldX + 256 - 32) * size / 256, y + (this.location.worldY + 256 - 32) * size / 256, 50)
         Renderer.rotate(rotation)
-
-        Renderer.colorize(0, 0, 0)
-        if (border) Renderer.drawRect(Renderer.BLACK, -w / 2 - 1, -h / 2 - 1, w + 2, h + 2)
+        
+        if (border != "none" || border == false){
+            this.updatePlayerColor()
+            Renderer.colorize(255,255,255)
+            Renderer.drawRect(Renderer.color(this.playerColor[0] ?? 0, this.playerColor[1] ?? 0, this.playerColor[2] ?? 0, this.playerColor[3] ?? 255), -w / 2 - 1, -h / 2 - 1, w + 2, h + 2)
+        }
 
         Renderer.colorize(255, 255, 255)